% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{print.summary.XDNUTS}
\alias{print.summary.XDNUTS}
\title{Function for printing an object of class summary.XDNUTS}
\usage{
\method{print}{summary.XDNUTS}(x, ..., digits = 3)
}
\arguments{
\item{x}{an object of class summary.XDNUTS}

\item{...}{additional values to pass. These currently do nothing.}

\item{digits}{number of digits for rounding the output. Default value is 3.}
}
\value{
No return value.
}
\description{
Print to console the statistics about the MCMC samples obtained with
a call to the function \link{summary.XDNUTS}. See that for details.
}
