% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{print.XDNUTS}
\alias{print.XDNUTS}
\title{Function for printing an object of class XDNUTS}
\usage{
\method{print}{XDNUTS}(x, ..., digits = 3, show_all = FALSE)
}
\arguments{
\item{x}{an object of class XDNUTS.}

\item{...}{additional arguments to pass. These currently do nothing.}

\item{digits}{a numeric scalar indicating the number of digits to show. Default value is 3.}

\item{show_all}{logical scalar indicating where to print all the summary statistics 
even if these are more than 10.}
}
\value{
Return a graphical object.
}
\description{
Print to console the specific of the algorithm used to generate an XDNUTS object.
}
